+++ aiohomematic/model/hub/data_point.py
@@ -103,6 +103,11 @@ class GenericHubDataPoint(CallbackDataPoint, PayloadMixin):
         return self._name_data.name
 
     @property
+    def signature(self) -> str:
+        """Return the signature of the data_point."""
+        return f"{self._category}/{self.name})"
+
+    @property
     def state_uncertain(self) -> bool:
         """Return, if the state is uncertain."""
         return self._state_uncertain